overhaul backend, move to esm
