+++ backend/walkSync.js
@@ -1,0 +1,14 @@
+import { readdirSync, statSync } from "node:fs";
+import { sep } from "node:path";
+
+export default function walkSync(dir, filelist = []) {
+	// add trailing slash if not present
+	if (!dir.endsWith(sep)) dir += sep;
+	for (const file of readdirSync(dir)) {
+		if (statSync(dir + file).isDirectory())
+			// read directories inside directories recursively
+			filelist = walkSync(dir + file + sep, filelist);
+		else filelist.push(dir + file);
+	}
+	return filelist;
+};
