+++ package.json
@@ -2,14 +2,15 @@
   "name": "website",
   "version": "1.0.0",
   "description": "The official website for the Faithful Project.",
-  "main": "index.js",
+  "main": "backend/index.js",
+  "type": "module",
   "scripts": {
     "ji": "bundle install",
     "jb": "bundle exec jekyll build",
     "start": "pnpm run ji && pnpm run jb && pm2 start \"node index.js\" --name \"WEBSITE\"",
     "dev": "concurrently \"npm:dev-*\"",
     "dev-watch-site": "bundle exec jekyll build --watch",
-    "dev-watch-app": "nodemon index.js --watch index.js --ignore '_site/**'"
+    "dev-watch-app": "nodemon backend/index.js --watch backend/ --ignore '_site/**'"
   },
   "repository": {
     "type": "git",
